Move NativeSeed into its own file
