+++ src/context/WalletProvider/config.ts
@@ -8,7 +8,7 @@ import { ShapeShiftVertical } from 'components/Icons/SSVerticalIcon'
 import { PinModal } from './KeepKey/PinModal'
 import { NativeImport } from './NativeWallet/NativeImport'
 import { NativePassword } from './NativeWallet/NativePassword'
-import { NativeSeed } from './NativeWallet/NativeSeed'
+import { NativeSeed } from './NativeWallet/NativeSeed/NativeSeed'
 import { NativeStart } from './NativeWallet/NativeStart'
 import { NativeSuccess } from './NativeWallet/NativeSuccess'
 import { NativeTestPhrase } from './NativeWallet/NativeTestPhrase/NativeTestPhrase'
